fix issues with multiple clients
