+++ scripts/_getDispatcher.js
@@ -1,5 +1,6 @@
 const {ethers} = require("hardhat");
 const fetchABI = require("./_fetchABI.js");
+const config = require("../config.json");
 
 const explorerOpUrl = "https://optimism-sepolia.blockscout.com/";
 const explorerBaseUrl = "https://base-sepolia.blockscout.com/";
@@ -13,17 +14,20 @@ async function getDispatcher (network) {
 
     let explorerUrl;
     let dispatcher;
+    let dispatcherAddress;
 
     if (network === "optimism") {
         explorerUrl = explorerOpUrl;
+        dispatcherAddress = config.proofsEnabled ? dispatcherAddress = process.env.OP_DISPATCHER : dispatcherAddress = process.env.OP_DISPATCHER_SIM;
 
-        const opDispatcherAbi = await fetchABI(explorerUrl, process.env.OP_DISPATCHER);
-        dispatcher = new ethers.Contract(process.env.OP_DISPATCHER, opDispatcherAbi, providerOptimism);
+        const opDispatcherAbi = await fetchABI(explorerUrl, dispatcherAddress);
+        dispatcher = new ethers.Contract(dispatcherAddress, opDispatcherAbi, providerOptimism);
     } else if (network === "base") {
         explorerUrl = explorerBaseUrl;
+        dispatcherAddress = config.proofsEnabled ? dispatcherAddress = process.env.BASE_DISPATCHER : dispatcherAddress = process.env.BASE_DISPATCHER_SIM;
 
-        const baseDispatcherAbi = await fetchABI(explorerUrl, process.env.BASE_DISPATCHER);
-        dispatcher = new ethers.Contract(process.env.BASE_DISPATCHER, baseDispatcherAbi, providerBase);
+        const baseDispatcherAbi = await fetchABI(explorerUrl, dispatcherAddress);
+        dispatcher = new ethers.Contract(dispatcherAddress, baseDispatcherAbi, providerBase);
     } else {
         throw new error(`Invalid network: ${network}`);
     }
