+++ scripts/_updateDispatcher.js
@@ -5,16 +5,22 @@
 // will compile your contracts, add the Hardhat Runtime Environment's members to the
 // global scope, and execute the script.
 const hre = require("hardhat");
+const path = require('path');
+const configRelativePath = process.env.CONFIG_PATH || 'config.json';
+const configPath = path.join(__dirname, '..' , configRelativePath);
 
 async function main() {
     const config = require(configPath);
+    const chanConfig = config["createChannel"];
+
+    const accounts = await hre.ethers.getSigners();
     const networkName = hre.network.name;
 
     let newDispatcher;
-    if (network === "optimism") {
-        newDispatcher = config.proofsEnabled ? process.env.OP_DISPATCHER_SIM : process.env.OP_DISPATCHER;
-    } else if (network === "base") {
-        newDispatcher = config.proofsEnabled ? process.env.BASE_DISPATCHER_SIM : process.env.BASE_DISPATCHER;
+    if (networkName === "optimism") {
+        newDispatcher = config.proofsEnabled === true ? process.env.OP_DISPATCHER_SIM : process.env.OP_DISPATCHER;
+    } else if (networkName === "base") {
+        newDispatcher = config.proofsEnabled === true ? process.env.BASE_DISPATCHER_SIM : process.env.BASE_DISPATCHER;
     } else {
         console.error("Invalid network name");
         process.exit(1);
