+++ scripts/_updateUcHandler.js
@@ -5,16 +5,22 @@
 // will compile your contracts, add the Hardhat Runtime Environment's members to the
 // global scope, and execute the script.
 const hre = require("hardhat");
+const path = require('path');
+const configRelativePath = process.env.CONFIG_PATH || 'config.json';
+const configPath = path.join(__dirname, '..' , configRelativePath);
 
 async function main() {
     const config = require(configPath);
+    const chanConfig = config["createChannel"];
+
+    const accounts = await hre.ethers.getSigners();
     const networkName = hre.network.name;
 
     let newUcHandler;
-    if (network === "optimism") {
-        newUcHandler = config.proofsEnabled ? process.env.OP_UC_MW_SIM : process.env.OP_UC_MW;
-    } else if (network === "base") {
-        newUcHandler = config.proofsEnabled ? process.env.BASE_UC_MW_SIM : process.env.BASE_UC_MW;
+    if (networkName === "optimism") {
+        newUcHandler = config.proofsEnabled === true ? process.env.OP_UC_MW_SIM : process.env.OP_UC_MW;
+    } else if (networkName === "base") {
+        newUcHandler = config.proofsEnabled === true ? process.env.BASE_UC_MW_SIM : process.env.BASE_UC_MW;
     } else {
         console.error("Invalid network name");
         process.exit(1);
